pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx-slide-copier"
version = "0.1.0"
description = "PowerPoint slide copier utility with deep copying of shapes, images, and formatting"
authors = [
    {name = "SpringMT", email = "today.is.sky.blue.sky@gmail.com"}
]
requires-python = ">=3.8"
dependencies = [
    "python-pptx>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
